Add superuser verification

+++ api/routes/syslogs.py
@@ -1,10 +1,10 @@
 from typing import Annotated
 
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from fastapi.responses import JSONResponse
 from api.models import SyslogIn, SyslogCreate, SyslogsOut, SyslogBase
 from core.db.engine import SessionDep
-from core.db.models import Syslog
+from core.db.models import Syslog, User
 from core.security import oauth2_scheme, validate_token, get_current_user
 
 router = APIRouter()
@@ -40,6 +40,14 @@ def create_syslog(session: SessionDep, syslog_in: SyslogCreate, token: Annotated
 @router.delete("")
 def delete_syslog(session: SessionDep, syslog_in: SyslogIn, token: Annotated[str, Depends(oauth2_scheme)]):
     validate_token(token)
+    user_name = get_current_user(token)
+    user = session.query(User).filter(User.username == user_name).first()
+    if not user:
+        raise HTTPException(status_code=403, detail="Access denied")
+
+    if not user.is_superuser:
+        raise HTTPException(status_code=403, detail="Forbidden : you are not a super user")
+
     # We add the filters only if they are not null
     filters = {k: v for k, v in syslog_in.model_dump().items() if v is not None}
     results = session.query(Syslog).filter_by(**filters).delete()
